+++ examples/full-page/src/pages/Home/tabs/ActivityTab/index.tsx
@@ -1,7 +1,7 @@
 import { Box, Inline, Link } from "@stripe/ui-extension-sdk/ui";
 import { DataTable } from "@stripe/ui-extension-sdk/ui/experimental";
 import { useRoute } from "@stripe/ui-extension-sdk/navigation";
-import { DATE_RANGE_OPTIONS, FilterSelect } from "@/components/FilterSelect";
+import { DATE_RANGE_OPTIONS, FilterSelect } from "@/pages/Home/components/FilterSelect";
 import { useActivityTab } from "./hooks/useActivityTab";
 
 interface ActivityTabProps {
